--- src/schoolOnline/courses/CourseServe.java
@@ -1,7 +1,0 @@
-package schoolOnline.courses;
-
-public class CourseServe {
-    public static Course courseCreate(){
-        return new Course();
-    }
-}
