+++ src/schoolOnline/entities/AdditionalMaterial.java
@@ -1,4 +1,4 @@
-package schoolOnline.lecture.additionalMaterial;
+package schoolOnline.entities;
 
 public class AdditionalMaterial {
     private int id;
