+++ src/schoolOnline/entities/Homework.java
@@ -1,4 +1,4 @@
-package schoolOnline.lecture.homework;
+package schoolOnline.entities;
 
 public class Homework {
     private int id;
