+++ src/schoolOnline/entities/Student.java
@@ -1,11 +1,11 @@
-package schoolOnline.student;
+package schoolOnline.entities;
 
 public class Student {
     public static int studentCount;
     private int id;
     String name;
     int age;
-    Student(){
+    public Student(){
         studentCount++;
     }
 }
