+++ src/schoolOnline/entitiesUtils/AdditionalMaterialUtil.java
@@ -1,0 +1,4 @@
+package schoolOnline.entitiesUtils;
+
+public class AdditionalMaterialUtil {
+}
