+++ src/schoolOnline/entitiesUtils/CourseUtil.java
@@ -1,0 +1,9 @@
+package schoolOnline.entitiesUtils;
+
+import schoolOnline.entities.Course;
+
+public class CourseUtil {
+    public static Course courseCreate(){
+        return new Course();
+    }
+}
