+++ src/schoolOnline/entitiesUtils/HomeworkUtil.java
@@ -1,0 +1,4 @@
+package schoolOnline.entitiesUtils;
+
+public class HomeworkUtil {
+}
