+++ src/schoolOnline/entitiesUtils/LectorUtil.java
@@ -1,0 +1,9 @@
+package schoolOnline.entitiesUtils;
+
+import schoolOnline.entities.Lector;
+
+public class LectorUtil {
+    public static Lector lectorCreate(){
+        return new Lector();
+    }
+}
