+++ src/schoolOnline/entitiesUtils/StudentUtil.java
@@ -1,0 +1,9 @@
+package schoolOnline.entitiesUtils;
+
+import schoolOnline.entities.Student;
+
+public class StudentUtil {
+    public static Student studentCreate(){
+        return new Student();
+    }
+}
