--- src/schoolOnline/lector/LectorServe.java
@@ -1,7 +1,0 @@
-package schoolOnline.lector;
-
-public class LectorServe {
-    public static Lector lectorCreate(){
-        return new Lector();
-    }
-}
