--- src/schoolOnline/lecture/LectureServe.java
@@ -1,10 +1,0 @@
-package schoolOnline.lecture;
-
-public class LectureServe {
-    public static Lecture lectureCreate(){
-        return new Lecture();
-    }
-    public static Lecture lectureCreate(int courseId){
-        return new Lecture(courseId);
-    }
-}
