--- src/schoolOnline/lecture/additionalMaterial/AdditionalMaterialServe.java
@@ -1,4 +1,0 @@
-package schoolOnline.lecture.additionalMaterial;
-
-public class AdditionalMaterialServe {
-}
