--- src/schoolOnline/lecture/homework/HomeworkServe.java
@@ -1,4 +1,0 @@
-package schoolOnline.lecture.homework;
-
-public class HomeworkServe {
-}
