--- src/schoolOnline/student/StudentServe.java
@@ -1,7 +1,0 @@
-package schoolOnline.student;
-
-public class StudentServe {
-    public static Student studentCreate(){
-        return new Student();
-    }
-}
